!B (CE-14658) preserve all 64 mask bits in MaterialProperties SetFlag() function.
Copied from Perforce
 Change: 1660525

+++ Code/Sandbox/Plugins/MaterialEditorPlugin/MaterialProperties.cpp
@@ -161,7 +161,7 @@ void SerializeWordToDegree(Serialization::IArchive& ar, uint16& word, char* name
 }
 
 template<typename T>
-void SetFlag(T& bitFieldInOut, int mask, bool set)
+void SetFlag(T& bitFieldInOut, uint64 mask, bool set)
 {
 	static_assert(std::is_integral<T>::value, "Type must be an integer type");
 	if (set)
@@ -1008,7 +1008,7 @@ void CMaterialSerializer::SerializeShaderParams(Serialization::IArchive& ar, boo
 			break;
 		}
 		default:
-			//Apparently unknown type parameters are a frequent occurence, do not warn
+			//Apparently unknown type parameters are a frequent occurrence, do not warn
 			//CryWarning(VALIDATOR_MODULE_EDITOR, VALIDATOR_WARNING, "Unknown shader parameter type found in material %s: %s", m_pMaterial->GetName().c_str(), param.m_Name);
 			break;
 		}
